pyproject.toml
[project]
name = "interop-router"
version = "0.1.1"
description = "Interoperate between AI providers using the OpenAI Responses API as a common interface."
readme = "README.md"
authors = [
    { name = "David Koleczek", email = "45405824+DavidKoleczek@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76,<1.0",
    "google-genai>=1.57,<2.0",
    "openai[aiohttp]>=2.15,<3.0",
    "pydantic>=2.12,<3.0",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1",
    "nbclient>=0.10",
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "ruff>=0.14",
    "ty>=0.0.12",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "RUF",  # ruff
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "PIE",  # flake8-pie
    "PT",   # flake-pytest-style
    "PTH",  # flake8-use-pathlib
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
]
ignore = [
    "E501",   # line-too-long (formatter makes best-effort attempt)
    # The following rules conflict with the formatter, see: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",   # tab-indentation
    "E111",   # indentation-with-invalid-multiple
    "E114",   # indentation-with-invalid-multiple-comment
    "E117",   # over-indented
    "ISC002", # multi-line-implicit-string-concatenation
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
asyncio_mode = "auto"
